requirements.txt
alembic==1.13.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.3
oauthlib==3.2.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
